LayoutTests/media/modern-media-controls/controls-visibility-support/controls-visibility-support-controls-on.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<video src="../../content/test.mp4" style="width: 320px; height: 240px;" controls></video>
<div id="shadow"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>ControlsVisibilitySupport</code> behavior without controls.");

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("video");
const mediaController = createControls(shadowRoot, media, null);

media.addEventListener("loadedmetadata", function() {
    debug("");
    debug("Media has loaded metadata");
    shouldBe("media.videoTracks.length", "1");
    shouldBeTrue("mediaController.controls.controlsBar.visible");
    shouldBeTrue("mediaController.controls.startButton.visible");
    debug("");
    shadowRoot.host.remove();
    media.remove();
    finishJSTest();
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
